bug(filenames):rename filenames  rename files to conform with airbnb naming styles  [Finishes #165379051]
