+++ Api/src/controllers/AccountController.js
@@ -1,4 +1,4 @@
-import AccountService from '../services/account.service';
+import AccountService from '../services/AccountService';
 
 class AccountController {
   static createAccount(req, res) {
