+++ Api/src/routes/accountRoute.js
@@ -1,5 +1,5 @@
 import { Router } from 'express';
-import AccountController from '../controllers/account.controller';
+import AccountController from '../controllers/AccountController';
 
 const router = Router();
