+++ Api/src/routes/userRoute.js
@@ -1,5 +1,5 @@
 import { Router } from 'express';
-import UserController from '../controllers/user.controller';
+import UserController from '../controllers/UserController';
 
 const router = Router();
